Surround delete image with existing check and try catch handler

+++ Source/Backend/vereinswebseite/data_cleanup.py
@@ -24,14 +24,17 @@ def delete_unused_images():
 
 
 def _delete_image(image_name):
-    fn = images.path(image_name)
-    os.remove(fn)
+    try:
+        fn = images.path(image_name)
+        if os.path.isfile(fn):
+            os.remove(fn)
+    except:
+        pass
 
 
 def _get_all_uploaded_image_filenames():
-    image_dir = os.path.dirname(images.path("doesnotmatter.png"))
-
     try:
+        image_dir = os.path.dirname(images.path("doesnotmatter.png"))
         filenames = [f for f in os.listdir(image_dir) if os.path.isfile(os.path.join(image_dir, f))]
         filenames = [f for f in filenames if f.endswith(IMAGES)]
         return filenames
